+++ cobaya/likelihoods/H0/freedman2020.py
@@ -2,4 +2,7 @@ from cobaya.likelihoods._base_classes import _H0_prototype
 
 
 class freedman2020(_H0_prototype):
+    r"""
+    Local $H_0$ measurement from \cite{Freedman:2020dne}.
+    """
     pass