+++ cobaya/likelihoods/bao/sdss_dr12_consensus_bao.py
@@ -2,5 +2,8 @@ from cobaya.likelihoods._base_classes import _bao_prototype
 
 
 class sdss_dr12_consensus_bao(_bao_prototype):
+    r"""
+    Likelihood of the power spectrum BAO of SDSS DR12 \cite{Alam:2016hwk}.
+    """
     pass