+++ cobaya/likelihoods/bao/sdss_dr12_consensus_full_shape.py
@@ -2,4 +2,7 @@ from cobaya.likelihoods._base_classes import _bao_prototype
 
 
 class sdss_dr12_consensus_full_shape(_bao_prototype):
+    r"""
+    Likelihood of the full-shape measurements of SDSS DR12 \cite{Alam:2016hwk}.
+    """
     pass