+++ cobaya/likelihoods/des_y1/galaxy_galaxy.py
@@ -2,4 +2,8 @@ from cobaya.likelihoods._base_classes import _des_prototype
 
 
 class galaxy_galaxy(_des_prototype):
-    pass
+    r"""
+    Galaxy-galaxy lensing data from the first year of the Dark Energy Survey (DES Y1)
+    \cite{Abbott:2017wau}.
+    """
+    bibtex_file = 'des_y1.bibtex'