+++ cobaya/likelihoods/planck_2015_plikHM_TT_unbinned/__init__.py
@@ -1,5 +1,9 @@
-from cobaya.likelihoods._base_classes import _planck_clik_prototype
+from cobaya.likelihoods._base_classes import Planck2015Clik
 
 
-class planck_2015_plikHM_TT_unbinned(_planck_clik_prototype):
+class planck_2015_plikHM_TT_unbinned(Planck2015Clik):
+    r"""
+    High-$\ell$ temperature-only likelihood (unbinned) of Planck's 2015 data release
+    \cite{Aghanim:2015xee}.
+    """
     pass