+++ cobaya/likelihoods/planck_2018_highl_plik/TTTEEE_lite.py
@@ -1,5 +1,9 @@
-from cobaya.likelihoods._base_classes import _planck_clik_prototype
+from cobaya.likelihoods._base_classes import Planck2018Clik
 
 
-class TTTEEE_lite(_planck_clik_prototype):
+class TTTEEE_lite(Planck2018Clik):
+    r"""
+    High-$\ell$ temperature and polarization, foreground-marginalized \textsc{plik\_lite}
+    likelihood of Planck's 2018 data release \cite{Aghanim:2019ame}.
+    """
     pass