+++ cobaya/likelihoods/planck_2018_highl_plik/TT_lite.py
@@ -1,5 +1,9 @@
-from cobaya.likelihoods._base_classes import _planck_clik_prototype
+from cobaya.likelihoods._base_classes import Planck2018Clik
 
 
-class TT_lite(_planck_clik_prototype):
+class TT_lite(Planck2018Clik):
+    r"""
+    High-$\ell$ temperature-only, foreground-marginalized \textsc{plik\_lite} likelihood
+    of Planck's 2018 data release \cite{Aghanim:2019ame}.
+    """
     pass