+++ cobaya/likelihoods/planck_2018_highl_plik/TT_lite_native.py
@@ -2,4 +2,9 @@ from cobaya.likelihoods._base_classes import _planck_pliklite_prototype
 
 
 class TT_lite_native(_planck_pliklite_prototype):
+    r"""
+    High-$\ell$ temperature-only, foreground-marginalized likelihood of Planck's 2018 data
+    release \cite{Aghanim:2019ame} (Python re-implementation of the official
+    \textsc{plik\_lite}, by E.~Calabrese, and A.~Lewis).
+    """
     pass