+++ cobaya/likelihoods/planck_2018_lensing/__init__.py
@@ -4,9 +4,21 @@ from cobaya.likelihoods._base_classes._planck_clik_prototype import \
 
 
 class native(_CMBlikes):
+    r"""
+    Lensing likelihood of Planck's 2018 data release based on temperature+polarization
+    map-based lensing reconstruction \cite{Aghanim:2018oex} (native Python
+    re-implementation by A.~Lewis).
+    """
     install_options = {"github_repository": "CobayaSampler/planck_supp_data_and_covmats",
                        "github_release": last_version_supp_data_and_covmats}
 
+    bibtex_file = 'PlanckLensing2018.bibtex'
+
 
 class CMBMarged(native):
+    r"""
+    Lensing likelihood of Planck's 2018 data release based on temperature+polarization
+    map-based lensing reconstruction, marginalized over the CMB power spectra
+    \cite{Aghanim:2018oex} (native Python re-implementation by A.~Lewis).
+    """
     pass