+++ cobaya/likelihoods/sn/jla_lite.py
@@ -2,4 +2,9 @@ from cobaya.likelihoods._base_classes import _sn_prototype
 
 
 class jla_lite(_sn_prototype):
+    r"""
+    Likelihood (marginalized over nuisance parameters) of the JLA type Ia supernova sample
+    \cite{Betoule:2014frx}, based on observations obtained by the SDSS-II and SNLS
+    collaborations.
+    """
     pass